unsplit and rename to realm.rb
