+++ lib/facter/realm.rb
@@ -1,6 +1,18 @@
-require_relative 'util/sssd'
-if defined? Facter::Util::Sssd
-  # == Fact: foreman_sssd
+require_relative 'util/realm'
+if defined? Facter::Util::Realm
+  # == Fact: foreman_ipa
+  Facter.add(:foreman_ipa, :type => :aggregate) do
+    {
+      :default_realm => 'global/realm',
+      :default_server => 'global/server',
+    }.each do |key, path|
+      chunk(key) do
+        val = Facter::Util::Realm.ipa_value(path)
+        {key => val} if val
+        # == Fact: foreman_sssd
+      end
+    end
+  end
   Facter.add(:foreman_sssd, :type => :aggregate) do
     {
       :services => 'target[.="sssd"]/services',
@@ -10,7 +22,7 @@ if defined? Facter::Util::Sssd
       :user_attributes => 'target[.="ifp"]/user_attributes',
     }.each do |key, path|
       chunk(key) do
-        val = Facter::Util::Sssd.sssd_value(path)
+        val = Facter::Util::Realm.sssd_value(path)
         {key => val} if val
       end
     end
